move getFinalNfps to clipper-flow
