+++ packages/geometry-utils/src/worker-flow/place-flow.ts
@@ -1,11 +1,11 @@
-import { get_result_wasm, get_placement_data_wasm, get_first_placement_wasm, init_place_content_debug_wasm, place_paths_wasm } from 'wasm-nesting';
+import { get_result_wasm, get_placement_data_wasm, get_first_placement_wasm, place_paths_wasm, apply_nfps_wasm, get_combined_nfps_wasm, get_final_nfp_wasm } from 'wasm-nesting';
 import type { Point, PolygonNode } from '../types';
 import ClipperWrapper from '../clipper-wrapper';
-import { serializePolygonNodes } from '../helpers';
+import { generateNFPCacheKey, serializePolygonNodes } from '../helpers';
 import PlaceContent from './place-content';
 import NFPWrapper from './nfp-wrapper';
 import { PointF32 } from '../geometry';
-
+import { f32 } from './types';
 
 function getPlacementData(
     finalNfp: Point<Int32Array>[][],
@@ -96,6 +96,60 @@ function getResult(placements: number[][], pathItems: number[][], fitness: numbe
     return result.buffer as ArrayBuffer;
 }
 
+function getFinalNfps(
+        placeContent: PlaceContent,
+        placed: PolygonNode[],
+        path: PolygonNode,
+        binNfp: NFPWrapper,
+        placement: f32[]
+    ): Point < Int32Array > [][] {
+    const tmpPoint: Point<Float32Array> = PointF32.create();
+    let i: number = 0;
+    let key: number = 0;
+
+    let totalNfps: Point<Int32Array>[][] = [];
+
+    for (i = 0; i < placed.length; ++i) {
+        key = generateNFPCacheKey(placeContent.rotations, false, placed[i], path);
+
+        if (!placeContent.nfpCache.has(key)) {
+            continue;
+        }
+
+        tmpPoint.fromMemSeg(placement, i);
+
+        // console.log(get_nfp_cache_value(key, new Uint8Array(placeContent.buffer)), new Float32Array(placeContent.nfpCache.get(key)));
+
+        totalNfps = totalNfps.concat(
+            ClipperWrapper.deserializePolygons(apply_nfps_wasm(new Float32Array(placeContent.nfpCache.get(key)), tmpPoint.x, tmpPoint.y))
+        );
+    }
+
+    const wasmRes = get_combined_nfps_wasm(
+        ClipperWrapper.serializePolygons(totalNfps)
+    );
+
+    const combinedNfp = ClipperWrapper.deserializePolygons(wasmRes);
+
+
+    if (combinedNfp.length === 0) {
+        return [];
+    }
+
+    // difference with bin polygon
+    const clipperBinNfp: Point<Int32Array>[][] = ClipperWrapper.nfpToClipper(binNfp);
+
+    const wasmResult = get_final_nfp_wasm(
+        ClipperWrapper.serializePolygons(combinedNfp),
+        ClipperWrapper.serializePolygons(clipperBinNfp)
+    )
+
+    const finalNfp = ClipperWrapper.deserializePolygons(wasmResult);
+
+
+    return finalNfp;
+}
+
 export function placePaths(buffer: ArrayBuffer): ArrayBuffer {
     const placeContent: PlaceContent = new PlaceContent().init(buffer);
     const firstPoint: Point<Float32Array> = PointF32.create();
@@ -145,7 +199,7 @@ export function placePaths(buffer: ArrayBuffer): ArrayBuffer {
                 continue;
             }
 
-            finalNfp = ClipperWrapper.getFinalNfps(placeContent, placed, node, binNfp, placement);
+            finalNfp = getFinalNfps(placeContent, placed, node, binNfp, placement);
 
             if (finalNfp.length === 0) {
                 continue;
